requirements.txt
Jinja2==3.1.2
python-dotenv==0.19.1
requests==2.26.0
plotly==5.3.1
pandas==1.3.3
numpy==1.21.2